refactor(elastic-farms): use new API
