+++ .env.production
@@ -36,3 +36,5 @@ REACT_APP_KS_SETTING_API=https://ks-setting.kyberswap.com/api
 REACT_APP_GTM_ID=GTM-TRQCJ8F
 
 REACT_APP_GOOGLE_RECAPTCHA_KEY=6LfgKRMiAAAAAPDTmXAM9LZLYAlE1zACfCJXup96
+
+REACT_APP_POOL_FARM_BASE_URL=https://pool-farm.kyberswap.com
